+++ app/routes/data-api/project/pattern_matchings.js
@@ -67,20 +67,25 @@ router.param('paging', function(req, res, next, paging){
 
 router.get('/:patternMatching/rois/:paging', function(req, res, next) {
     res.type('json');
-    model.patternMatchings.getRoisForId({
-        patternMatchingId: req.params.patternMatching,
-        wherePresent: req.query.search == 'present',
-        whereNotPresent: req.query.search == 'not_present',
-        whereUnvalidated: req.query.search == 'unvalidated',
-        bestPerSite: req.query.search == 'best_per_site',
-        bestPerSiteDay: req.query.search == 'best_per_site_day',
-        byScorePerSite: req.query.search == 'by_score_per_site',
-        byScore: req.query.search == 'by_score',
-        limit: req.paging.limit || 100,
-        offset: req.paging.offset || 0,
-    }).then(function(rois) {
-        res.json(rois);
-    }).catch(next);
+    let prom
+    if (req.query.search == 'best_per_site' || req.query.search == 'best_per_site_day') {
+        prom = model.patternMatchings.getRoisBestPerSiteForId(req.params.patternMatching, req.query.search == 'best_per_site_day')
+    } else {
+        prom = model.patternMatchings.getRoisForId({
+            patternMatchingId: req.params.patternMatching,
+            wherePresent: req.query.search == 'present',
+            whereNotPresent: req.query.search == 'not_present',
+            whereUnvalidated: req.query.search == 'unvalidated',
+            byScorePerSite: req.query.search == 'by_score_per_site',
+            byScore: req.query.search == 'by_score',
+            limit: req.paging.limit || 100,
+            offset: req.paging.offset || 0,
+        })
+    }
+    prom.then(function(rois) {
+            res.json(rois);
+        })
+        .catch(next);
 });
 
 router.get('/:patternMatching/site-index', function(req, res, next) {